CVSL-2584 AFER | Preparation for displaying - code issue from review
